# Implement array view elision