--- a/compiler/optimizations/Makefile.share
+++ b/compiler/optimizations/Makefile.share
@@ -17,6 +17,7 @@
 # limitations under the License.
 
 OPTIMIZATIONS_SRCS = \
+	arrayViewElision.cpp \
 	bulkCopyRecords.cpp \
 	copyPropagation.cpp \
 	deadCodeElimination.cpp \