--- a/test/gpu/native/optimizations/arrayViewElision/gpuViewElision.compopts
+++ b/test/gpu/native/optimizations/arrayViewElision/gpuViewElision.compopts
@@ -1,0 +1,1 @@
+-sdebugBulkTransfer -sdebugShortArrayTransfer