… compiler/optimizations/CMakeLists.txt
… compiler/resolution/preFold.cpp
… frontend/include/chpl/uast/PragmaList.h
… frontend/include/chpl/uast/prim-ops-list.h
… frontend/lib/resolution/prims.cpp
… man/chpl.rst
… modules/internal/ChapelArray.chpl
… modules/internal/ChapelArrayViewElision.chpl
… modules/internal/ChapelBase.chpl
… modules/internal/ChapelDomain.chpl
… modules/internal/ChapelShortArrayTransfer.chpl
… modules/internal/ChapelStandard.chpl
… modules/internal/ChapelTuple.chpl
… modules/standard/CTypes.chpl
… runtime/include/localeModels/flat/chpl-locale-model.h
… runtime/include/localeModels/gpu/chpl-locale-model.h
… test/compflags/bradc/help/userhelp.good
… test/compflags/ferguson/print-module-resolution.good
… test/gpu/native/optimizations/arrayViewElision/gpuViewElision.chpl
… test/gpu/native/optimizations/arrayViewElision/gpuViewElision.execopts
… test/gpu/native/optimizations/arrayViewElision/gpuViewElision.good
… test/optimizations/arrayViewElision/avoidExtraCalls.chpl
… test/optimizations/arrayViewElision/avoidExtraCalls.comm-none.good
… test/optimizations/arrayViewElision/avoidExtraCalls.compopts
… test/optimizations/arrayViewElision/avoidExtraCalls.good
… test/optimizations/arrayViewElision/basic.chpl
… test/optimizations/arrayViewElision/basic.compopts
… test/optimizations/arrayViewElision/basic.good
… test/optimizations/arrayViewElision/constChecking-error.good
… test/optimizations/arrayViewElision/constChecking-no-error.good