@e-kayrakli 2024-02-10
Here's a quick performance test based on a test this PR adds https://github.com/chapel-lang/chapel/pull/24390/files#diff-87b50ddd74dbe56778cc9659cf389dd42231fbf66147dd94487a1a37b0e63183


This shows the throughput of repeatedly copying slices of given size. Performance numbers are GB/s. Last two columns are speedup.


elems/copy | loop | slice   (`main`) | slice   (PR) | x over "loop" | x over `main`
-- | -- | -- | -- | -- | --
10 | 32.01 | 0.91 | 26.55 | 0.83 | 29.18
100 | 52.78 | 9.44 | 78.47 | 1.49 | 8.31
1000 | 73.71 | 48.74 | 87.22 | 1.18 | 1.79
10k | 42.72 | 50.24 | 52.8 | 1.24 | 1.05
10m | 6.82 | 9.14 | 9.15 | 1.34 | 1.00

So, in really pathological cases (copying slices of size 10), this PR improves performance by more than 29x. In such scenarios, slices perform about 80% of looping over data. Note that for these sizes, cut off where slicing outperforms looping was 10k. After this PR it'll be 100.

For the pathological case: it is the only case where short array transfer optimization helps (my current default is size 50). Hacking my way around, not doing dynamic checks brings performance on-par with looping. So, I think I squeezed everything I could without adding more significant code analysis in the compiler. This dynamic check involves checking the size of the slice in local compilations. In distributed or GPU compilations we'll also check the locale of both sides.