@jeremiah-corrado 2024-07-01
On this branch I'm seeing some performance improvements for a poisson solver code that I'm writing a blog article about. The code is a finite difference solver that applies the following boundary condition procedure to a 2D array every iteration:

```chapel
proc neumannBC(ref p: [] real) {
    p[0,    ..] = p[1, ..];    // dp/dy = 0 @ y = 0
    p[ny-1, ..] = p[ny-2, ..]; // dp/dy = 0 @ y = yLen
}
```

Compiling the code on this branch with `--report-array-view-elision`, produces the following output:

<details>

```
ArrayViewElision supported nsPoisson.chpl:30
  lhsBaseType: [domain(2,int(64),one)] real(64)
  lhsIndexingExprs:
    int(64)
    range(int(64),neither,one)
  rhsBaseType: [domain(2,int(64),one)] real(64)
  rhsIndexingExprs:
    int(64)
    range(int(64),neither,one)

ArrayViewElision supported nsPoisson.chpl:31
  lhsBaseType: [domain(2,int(64),one)] real(64)
  lhsIndexingExprs:
    int(64)
    range(int(64),neither,one)
  rhsBaseType: [domain(2,int(64),one)] real(64)
  rhsIndexingExprs:
    int(64)
    range(int(64),neither,one)

ArrayViewElision supported nsPoisson.chpl:90
  lhsBaseType: _ref(_array(unmanaged [domain(2,int(64),one)] real(64)))
  lhsIndexingExprs:
    int(64)
    range(int(64),neither,one)
  rhsBaseType: _ref(_array(unmanaged [domain(2,int(64),one)] real(64)))
  rhsIndexingExprs:
    int(64)
    range(int(64),neither,one)

ArrayViewElision supported nsPoisson.chpl:91
  lhsBaseType: _ref(_array(unmanaged [domain(2,int(64),one)] real(64)))
  lhsIndexingExprs:
    int(64)
    range(int(64),neither,one)
  rhsBaseType: _ref(_array(unmanaged [domain(2,int(64),one)] real(64)))
  rhsIndexingExprs:
    int(64)
    range(int(64),neither,one)
```

where, lines 90 and 91 correspond the above procedure, and lines 30 and 31 are part of some setup code that is only called once.

</details>

Running the solver for 10,000 iterations with various problem sizes, both on this branch and on main produced the following average runtimes (over 10 runs).  Here, the size of the copies is `n`, but the total problem size is `n*n`.

| n | main (ms) | branch (ms) | % speedup |
| :--- | :--- | :--- | :--- |
| 256 | 242.3 | 198.2 | 18.2 |
| 512 | 250.8 | 234.9 | 6.3 |
| 1024 | 430.6 | 417.3 | 3.1 |
| 2048 | 1429 | 1433 |  -0.2 |

Results were collected on a system with a 32-core AMD EPYC.

I'm not too surprised to see that the speedup is entirely amortized for the larger problem sizes because the main computation is over the entire domain (`n*n` elements), while the boundary condition computation only copies `2*n` elements. Regardless, I think this clearly shows the benefits of the optimization (especially for programs whose runtime is more dominated by slice assignments than this one).