@e-kayrakli 2024-07-01
@riftEmber -- Hi Anna, I added you as a reviewer particulary for changes to some `c_ptr*` functions. These are performance optimizations for code that is supposed to be fast. It avoids some costly operations by opting into lower-level ways of doing them and hiding some more expensive checks under `boundsChecking`