@e-kayrakli 2024-07-03
@mppf / others -- Note that https://github.com/chapel-lang/chapel/pull/24390/commits/82b41f8b5521f0693fd34a5022ebbfe74cbd6db7 tightens some comm count tests. It looks like the refactor on DR iterators in this PR reduces comm counts when iterating over a remote DR. That refactor added a special-case for 1D DRs where we use a range instead of a full blown domain in the iterator.