@e-kayrakli 2024-07-03
@benharsh re short array transfer (SAT) threshold: It looks like there's some room upwards on a workstation and a single-socket EX node. But running on a dual-socket XC, I see that we are close to the margin. That being said, I see that there's a drop in benefits after 60 elements/transfer. The current threshold is 50 elements. I'm inclined to bump it modestly to 60.

There could be some more room for some applications/hardware. The tables below could help seeing the limits of SAT.

\* marks break-even points

### Workstation
- Intel(R) Core(TM) i5-11400 @ 2.60GHz
- LLVM 15


num elems | GB/s w/ SAT | GB/s w/o SAT | SAT performance
-- | -- | -- | --
20 | 44.01 | 24.72 | 1.78
40 | 57.52 | 41.78 | 1.38
60 | 66.55 | 59.12 | 1.13
*80 | 72.01 | 71.26 | 1.01
*100 | 82.16 | 80.84 | 1.02
*120 | 82.36 | 79.84 | 1.03
140 | 79.81 | 91.57 | 0.87
160 | 81.17 | 96.43 | 0.84
180 | 84.55 | 98.24 | 0.86
200 | 83.25 | 97.95 | 0.85
220 | 83.82 | 101.74 | 0.82
240 | 82.5 | 105.97 | 0.78
260 | 79.37 | 103.64 | 0.77
280 | 76.78 | 112.53 | 0.68
300 | 76.3 | 114.06 | 0.67

## Single-socket EX node
- EPYC 7763 64-Core 2.45 GHz Milan
- LLVM 18



num elems | GB/s w/ SAT | GB/s w/o SAT | SAT performance
-- | -- | -- | --
20 | 45.09 | 17.66 | 2.55
40 | 55.71 | 29.89 | 1.86
60 | 58.6 | 39.73 | 1.47
80 | 60.75 | 51.5 | 1.18
100 | 74.05 | 56.45 | 1.31
*120 | 72.24 | 63.09 | 1.15
*140 | 67.81 | 68.99 | 0.98
160 | 68.36 | 75.32 | 0.91
180 | 77.59 | 80.44 | 0.96
200 | 78.58 | 80.78 | 0.97
220 | 78.11 | 83.78 | 0.93
240 | 79.22 | 87.35 | 0.91
260 | 79 | 81.38 | 0.97
280 | 76.89 | 79.05 | 0.97
300 | 77.01 | 78.57 | 0.98

## Dual-socket XC node
- Xeon 8160 CPU @ 2.10GHz SKL
- LLVM 14


num elems | GB/s w/ SAT | GB/s w/o SAT | SAT performance
-- | -- | -- | --
20 | 22.83 | 13.47 | 1.69
40 | 35.35 | 22.66 | 1.56
60 | 44.41 | 28.71 | 1.55
*80 | 40.36 | 39.12 | 1.03
100 | 32.61 | 44.79 | 0.73
120 | 39 | 45.19 | 0.86
140 | 44.8 | 51.4 | 0.87
160 | 45.33 | 46.53 | 0.97
180 | 48.22 | 52.47 | 0.92
200 | 48.07 | 55.11 | 0.87
220 | 48.02 | 59.38 | 0.81
240 | 45.57 | 62.01 | 0.73
260 | 46.53 | 65.55 | 0.71
280 | 48.44 | 65.24 | 0.74
300 | 46.59 | 67.29 | 0.69