@e-kayrakli 2024-07-06
All testing is clean. I updated https://github.com/chapel-lang/chapel/pull/24390#issuecomment-1936776679 with the most recent performance numbers. I'll merge this next week.