@mppf 2024-07-08
Great! I'm looking forward to having this change.

When you get to writing the final PR message / merging, could you make sure to update/comment on issues #16133 and #24343? In particular, can #16133 be closed?