@e-kayrakli 2024-07-08
> When you get to writing the final PR message / merging, could you make sure to update/comment on issues https://github.com/chapel-lang/chapel/issues/16133 and https://github.com/chapel-lang/chapel/issues/24343? In particular, can https://github.com/chapel-lang/chapel/issues/16133 be closed?

I posted under those issues (neither can be closed via this PR) for some next steps after this PR. Also back-linked those comments in the PR message.